app/requirements.txt
Django==3.1
djangorestframework==3.11
psycopg2-binary==2.8.5
pandas==1.0.3
natsort==7.0.1
gunicorn>=20.0.4,<20.1
django-cors-headers==3.6.0

pytest==6.2.4
pytest-django==4.2.0
pytest-snapshot==0.6.1
django-extensions

flake8==3.8.3
PyYAML==5.4.1